+++ src/hyperbase/readers/__init__.py
@@ -1,7 +1,5 @@
 from hyperbase.readers import (
     txt,  # noqa: F401
-    url,  # noqa: F401
-    wikipedia,  # noqa: F401
 )
 from hyperbase.readers.reader import Reader, get_reader, list_readers, register_reader